[MDH-109] fix : 충돌 해결
