+++ common/src/main/java/com/mdh/common/global/config/JpaConfig.java
@@ -1,9 +1,6 @@
 package com.mdh.common.global.config;
 
-import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.annotation.PostConstruct;
-import jakarta.persistence.EntityManager;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
@@ -16,10 +13,5 @@ public class JpaConfig {
     @PostConstruct
     public void setTimeZone() {
         TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
-    }
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory(EntityManager em) {
-        return new JPAQueryFactory(em);
     }
 }
