+++ user/src/main/java/com/mdh/user/shop/presentation/ShopController.java
@@ -2,14 +2,20 @@ package com.mdh.user.shop.presentation;
 
 import com.mdh.user.global.ApiResponse;
 import com.mdh.user.shop.application.ShopService;
+import com.mdh.user.shop.application.dto.ReservationShopSearchResponse;
 import com.mdh.user.shop.application.dto.ShopDetailInfoResponse;
 import com.mdh.user.shop.application.dto.ShopResponses;
+import com.mdh.user.shop.presentation.dto.ReservationShopSearchRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Map;
 
 @RequiredArgsConstructor
 @RestController
@@ -25,6 +31,14 @@ public class ShopController {
 
         var result = shopService.findByConditionWithWaiting(multiParams, pageable);
         return new ResponseEntity<>(ApiResponse.ok(result), HttpStatus.OK);
+    }
+
+    @GetMapping("/reservations")
+    public ResponseEntity<ApiResponse<ReservationShopSearchResponse>> searchReservationShops(
+            @PageableDefault(sort = "createdDate", direction = Sort.Direction.ASC) Pageable pageable,
+            @RequestParam Map<String, String> params) {
+        var reservationShopSearchResponse = shopService.searchReservationShop(pageable, ReservationShopSearchRequest.of(params));
+        return ResponseEntity.ok(ApiResponse.ok(reservationShopSearchResponse));
     }
 
     @GetMapping("/{shopId}")
