+++ user/src/test/java/com/mdh/user/shop/application/ShopServiceTest.java
@@ -5,6 +5,7 @@ import com.mdh.common.shop.persistence.ShopRepository;
 import com.mdh.common.shop.persistence.dto.ReservationShopSearchQueryDto;
 import com.mdh.user.DataInitializerFactory;
 import com.mdh.user.shop.application.dto.ShopDetailInfoResponse;
+import com.mdh.user.shop.presentation.dto.ReservationShopSearchRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
